More cleanup

+++ test/priqueue_tests.cpp
@@ -4,7 +4,7 @@
 #include <vector>
 
 struct event {
-    int value;
+    size_t value;
     void* pqn;
 };
 struct eventCompare
@@ -21,15 +21,15 @@ TEST(PriQueueTests, TestPushPop)
     size_t count = 20000;
 
     std::vector<event*> events;
-    for (int i = 0; i < count; i++) {
-        event* e = new event({rand(), nullptr});
+    for (size_t i = 0; i < count; i++) {
+        event* e = new event({(size_t)rand(), nullptr});
         events.push_back(e);
         pq.push(e);
     }
     EXPECT_EQ(pq.audit(), count);
 
     int last = pq.top()->value;
-    for (int i = 0; i < count; i++) {
+    for (size_t i = 0; i < count; i++) {
         int curr = pq.top()->value;
         EXPECT_GE(curr, last);
         last = curr;
@@ -37,7 +37,7 @@ TEST(PriQueueTests, TestPushPop)
     }
     EXPECT_EQ(pq.empty(), true);
 
-    for (int i = 0; i < count; i++) {
+    for (size_t i = 0; i < count; i++) {
         delete events[i];
     }
 }
@@ -48,8 +48,8 @@ TEST(PriQueueTests, TestPushPop2)
     size_t count = 4000;
 
     std::vector<event*> events;
-    for (int i = 0; i < count; i++) {
-        event* e = new event({rand(), nullptr});
+    for (size_t i = 0; i < count; i++) {
+        event* e = new event({(size_t)rand(), nullptr});
         events.push_back(e);
         pq.push(e);
         EXPECT_EQ(pq.audit(), i+1);
@@ -57,7 +57,7 @@ TEST(PriQueueTests, TestPushPop2)
     EXPECT_EQ(pq.audit(), count);
 
     int last = pq.top()->value;
-    for (int i = 0; i < count; i++) {
+    for (size_t i = 0; i < count; i++) {
         int curr = pq.top()->value;
         EXPECT_GE(curr, last);
         last = curr;
@@ -66,7 +66,7 @@ TEST(PriQueueTests, TestPushPop2)
     }
     EXPECT_EQ(pq.empty(), true);
 
-    for (int i = 0; i < count; i++) {
+    for (size_t i = 0; i < count; i++) {
         delete events[i];
     }
 }
@@ -77,7 +77,7 @@ TEST(PriQueueTests, TestErase)
     size_t count = 20000;
 
     std::vector<event*> events;
-    for (int i = 0; i < count; i++) {
+    for (size_t i = 0; i < count; i++) {
         event* e = new event({i, nullptr});
         events.push_back(e);
         pq.push(e);
@@ -85,18 +85,18 @@ TEST(PriQueueTests, TestErase)
     EXPECT_EQ(pq.audit(), count);
 
     // erase odd elements
-    for (int i = 1; i < count; i+=2) {
+    for (size_t i = 1; i < count; i+=2) {
         pq.erase(events[i]);
     }
     EXPECT_EQ(pq.audit(), count/2);
 
-    for (int i = 0; i < (count+1)/2; i++) {
-        EXPECT_GE(pq.top()->value % 2, 0);
+    for (size_t i = 0; i < (count+1)/2; i++) {
+        EXPECT_GE(pq.top()->value % 2, (size_t)0);
         pq.pop();
     }
     EXPECT_EQ(pq.empty(), true);
 
-    for (int i = 0; i < count; i++) {
+    for (size_t i = 0; i < count; i++) {
         delete events[i];
     }
 }
@@ -108,20 +108,20 @@ TEST(PriQueueTests, TestErase2)
 
     std::vector<event*> events;
     for (int i = count-1; i >= 0; i--) {
-        event* e = new event({i, nullptr});
+        event* e = new event({(size_t)i, nullptr});
         events.push_back(e);
         pq.push(e);
     }
     EXPECT_EQ(pq.audit(), count);
 
     // erase from front of queue
-    for (int i = 0; i < count; i++) {
+    for (size_t i = 0; i < count; i++) {
         pq.erase(events[i]);
     }
-    EXPECT_EQ(pq.audit(), 0);
+    EXPECT_EQ(pq.audit(), (size_t)0);
     EXPECT_EQ(pq.empty(), true);
 
-    for (int i = 0; i < count; i++) {
+    for (size_t i = 0; i < count; i++) {
         delete events[i];
     }
 }
